backend/functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
